migrate validation functions into a utils file for better organization
