+++ advisor-ui/src/pages/Signup/Signup.jsx
@@ -4,6 +4,14 @@ import { useNavigate, Link } from "react-router-dom";
 import { UserContext } from "../../UserContext.js";
 import { Button } from "@mui/material";
 import apiBase from "../../utils/apiBase";
+import {
+  validateNonEmptyFields,
+  validateEmail,
+  validatePassword,
+  validateYear,
+  validateCity,
+  validateState,
+} from "../../utils/signupValidationUtils";
 import "./Signup.css";
 
 export default function Signup() {
@@ -50,127 +58,37 @@ export default function Signup() {
     setAddress(e.target.value);
   };
 
-  const validateNonEmptyFields = () => {
-    if (
-      firstname === "" ||
-      lastname === "" ||
-      username === "" ||
-      email === "" ||
-      password === "" ||
-      year === "" ||
-      city === "" ||
-      locationState === "" ||
-      address === ""
-    ) {
-      return false;
-    }
-    return true;
-  };
-
-  const validateEmail = () => {
-    // Check if the email contains an '@' symbol
-    const atIndex = email.indexOf("@");
-    if (atIndex === -1) {
-      return false;
-    }
-
-    // Check if the email contains a domain (characters after the '@' symbol)
-    const domain = email.slice(atIndex + 1);
-    if (!domain) {
-      return false;
-    }
-
-    // Check if the domain contains a period
-    const periodIndex = domain.indexOf(".");
-    if (periodIndex === -1) {
-      return false;
-    }
-
-    // Check if the domain has at least one character before the period
-    const domainName = domain.slice(0, periodIndex);
-    if (!domainName) {
-      return false;
-    }
-
-    // Check if the domain has at least two characters after the period
-    const topLevelDomain = domain.slice(periodIndex + 1);
-    if (!topLevelDomain || topLevelDomain.length < 2) {
-      return false;
-    }
-
-    // All checks passed, email is valid
-    return true;
-  };
-
-  const validatePassword = () => {
-    if (password.length < 8) return false;
-    const specialCharacters = new Set(["!", "@", "#", "$", "%", "^", "&", "*"]);
-    const upperCaseA_Ascii = 65;
-    const upperCaseZ_Ascii = 90;
-    // validate presence of uppercase letter and special character
-    let hasUpperCase = false;
-    let hasSpecialCharacter = false;
-    for (let i = 0; i < password.length; i++) {
-      // if the character @ idx i's ascii value is within this range, it must be a capital letter
-      if (
-        password.charCodeAt(i) >= upperCaseA_Ascii &&
-        password.charCodeAt(i) <= upperCaseZ_Ascii
-      ) {
-        hasUpperCase = true;
-      }
-      if (specialCharacters.has(password[i])) {
-        hasSpecialCharacter = true;
-      }
-    }
-
-    return hasSpecialCharacter && hasUpperCase;
-  };
-
-  const validateYear = () => {
-    // regex test ensures year only contains digit and parseInt restricts min value of year
-    if (!/^\d+$/.test(year) || parseInt(year) <= 0) {
-      return false;
-    }
-    return true;
-  };
-
-  const validateCity = () => {
-    if (/\d/.test(city)) {
-      return false;
-    }
-    return true;
-  };
-
-  const validateState = () => {
-    const commonStateAbbreviatonLength = 2;
-    if (
-      /\d/.test(locationState) ||
-      locationState.length <= commonStateAbbreviatonLength
-    ) {
-      return false;
-    }
-    return true;
-  };
-
   const handleCreateAccount = async () => {
-    if (!validateNonEmptyFields()) {
+    if (
+      !validateNonEmptyFields(
+        firstname,
+        lastname,
+        username,
+        email,
+        password,
+        year,
+        city,
+        locationState,
+        address
+      )
+    ) {
       alert("Please enter all fields");
       return;
-    } else if (!validateEmail()) {
+    } else if (!validateEmail(email)) {
       alert("Please enter a valid email");
       return;
-    } else if (!validatePassword()) {
+    } else if (!validatePassword(password)) {
       alert(
         "Password must be at least 8 characters and contain one uppercase and special character"
       );
       return;
-    } else if (!validateYear()) {
+    } else if (!validateYear(year)) {
       alert("Year must only contain number and must be greater than 0");
       return;
-    } else if (!validateCity()) {
+    } else if (!validateCity(city)) {
       alert("City field should not contain any numbers");
       return;
-    } else if (!validateState()) {
+    } else if (!validateState(locationState)) {
       alert(
         "State field should contain the full state name and should not contain any numbers"
       );
